Moved JediAtoms to jediatoms.py
